move models from lib folder
