+++ src/components/funds/list/fund-card.tsx
@@ -1,6 +1,6 @@
 import StatusChip from '@/components/shared/chip/status-chip';
 import { ProgressBar } from '@/components/shared/progress/progress-bar';
-import { FundStatus } from '@/lib/models/funds/fund.model';
+import { FundStatus } from '@/models/funds/fund.model';
 import { ChipStatus } from '@/models/status/chip-status.enum';
 import { formatCurrency, formatDate } from '@/utils/format.utils';
