--- src/lib/models/userData.model.ts
@@ -1,8 +1,0 @@
-import { FundDetails } from './funds/fund.model';
-import { PlanDetails } from './funds/plan.model';
-
-export interface UserData {
-  id: string;
-  funds?: FundDetails[];
-  plans?: PlanDetails[];
-}
