+++ src/models/funds/group.model.ts
