+++ src/models/funds/plan.model.ts
@@ -1,4 +1,4 @@
-import { TransactionDetails } from './transaction.model';
+import { TransactionDetails } from '@/models/funds/transaction.model';
 
 export type PlanStatus = 'pending' | 'completed' | 'canceled';
