+++ src/models/user/userData.model.ts
@@ -1,0 +1,8 @@
+import { FundDetails } from '@/models/funds/fund.model';
+import { PlanDetails } from '@/models/funds/plan.model';
+
+export interface UserData {
+  id: string;
+  funds?: FundDetails[];
+  plans?: PlanDetails[];
+}
